+++ sdk/storage/Azure.Storage.Blobs.ChangeFeed/api/Azure.Storage.Blobs.ChangeFeed.net8.0.cs
@@ -55,6 +55,7 @@ namespace Azure.Storage.Blobs.ChangeFeed
         public Azure.Storage.Blobs.ChangeFeed.ChangeFeedEventPreviousInfo PreviousInfo { get { throw null; } }
         public bool? Recursive { get { throw null; } }
         public System.Guid RequestId { get { throw null; } }
+        public string RestoredContainerVersion { get { throw null; } }
         public string Sequencer { get { throw null; } }
         public string Snapshot { get { throw null; } }
         public System.Uri SourceUri { get { throw null; } }
@@ -74,6 +75,9 @@ namespace Azure.Storage.Blobs.ChangeFeed
         public static Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventType BlobPropertiesUpdated { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventType BlobSnapshotCreated { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventType BlobTierChanged { get { throw null; } }
+        public static Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventType ContainerCreated { get { throw null; } }
+        public static Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventType ContainerDeleted { get { throw null; } }
+        public static Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventType ContainerPropertiesUpdated { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventType Control { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventType RestorePointMarkerCreated { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventType UnspecifiedEventType { get { throw null; } }
@@ -93,7 +97,7 @@ namespace Azure.Storage.Blobs.ChangeFeed
     {
         public static Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEvent BlobChangeFeedEvent(string topic, string subject, Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventType eventType, System.DateTimeOffset eventTime, System.Guid id, Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventData eventData, long dataVersion, string metadataVersion) { throw null; }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationResult BlobChangeFeedEventAsyncOperationInfo(Azure.Storage.Blobs.Models.AccessTier? destinationAccessTier, bool wasAsyncOperation, string copyId) { throw null; }
-        public static Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventData BlobChangeFeedEventData(string blobOperationName, string clientRequestId, System.Guid requestId, Azure.ETag eTag, string contentType, long contentLength, Azure.Storage.Blobs.Models.BlobType blobType, string blobVersion, string containerVersion, Azure.Storage.Blobs.Models.AccessTier? blobAccessTier, long contentOffset, System.Uri destinationUri, System.Uri sourceUri, System.Uri uri, bool recursive, string sequencer, Azure.Storage.Blobs.ChangeFeed.ChangeFeedEventPreviousInfo previousInfo, string snapshot, System.Collections.Generic.Dictionary<string, Azure.Storage.Blobs.ChangeFeed.BlobPropertyChange> updatedBlobProperties, Azure.Storage.Blobs.ChangeFeed.BlobOperationResult asyncOperationInfo, Azure.Storage.Blobs.ChangeFeed.BlobTagsChange updatedBlobTags) { throw null; }
+        public static Azure.Storage.Blobs.ChangeFeed.BlobChangeFeedEventData BlobChangeFeedEventData(string blobOperationName, string clientRequestId, System.Guid requestId, Azure.ETag eTag, string contentType, long contentLength, Azure.Storage.Blobs.Models.BlobType blobType, string blobVersion, string containerVersion, Azure.Storage.Blobs.Models.AccessTier? blobAccessTier, long contentOffset, System.Uri destinationUri, System.Uri sourceUri, System.Uri uri, bool recursive, string sequencer, Azure.Storage.Blobs.ChangeFeed.ChangeFeedEventPreviousInfo previousInfo, string snapshot, System.Collections.Generic.Dictionary<string, Azure.Storage.Blobs.ChangeFeed.BlobPropertyChange> updatedBlobProperties, Azure.Storage.Blobs.ChangeFeed.BlobOperationResult asyncOperationInfo, Azure.Storage.Blobs.ChangeFeed.BlobTagsChange updatedBlobTags, string restoredContainerVersion = null) { throw null; }
         public static Azure.Storage.Blobs.ChangeFeed.BlobPropertyChange BlobChangeFeedEventUpdatedBlobProperty(string propertyName, string previousValue, string newValue) { throw null; }
         public static Azure.Storage.Blobs.ChangeFeed.BlobTagsChange BlobChangeFeedEventUpdatedBlobTags(System.Collections.Generic.Dictionary<string, string> previousTags, System.Collections.Generic.Dictionary<string, string> newTags) { throw null; }
         public static Azure.Storage.Blobs.ChangeFeed.ChangeFeedEventPreviousInfo ChangeFeedEventPreviousInfo(string softDeleteSnapshot, bool wasBlobSoftDeleted, string blobVersion, string lastVersion, Azure.Storage.Blobs.Models.AccessTier? previousTier) { throw null; }
@@ -107,14 +111,18 @@ namespace Azure.Storage.Blobs.ChangeFeed
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName AbortCopyBlob { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName ControlEvent { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName CopyBlob { get { throw null; } }
+        public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName CreateContainer { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName CreateRestorePointMarker { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName DeleteBlob { get { throw null; } }
+        public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName DeleteContainer { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName PutBlob { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName PutBlockList { get { throw null; } }
+        public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName RestoreContainer { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName SetBlobMetadata { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName SetBlobProperties { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName SetBlobTags { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName SetBlobTier { get { throw null; } }
+        public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName SetContainerMetadata { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName SnapshotBlob { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName UndeleteBlob { get { throw null; } }
         public static Azure.Storage.Blobs.ChangeFeed.BlobOperationName UnspecifiedApi { get { throw null; } }